api/requirements.txt
Flask==3.0.2
groq==0.4.2
python-dotenv==1.0.1
flask-cors==4.0.0
